@@ -4,7 +4,7 @@ version: 0.1.0-dev
 repository: https://github.com/verygoodtechnologies/cupertino_timer
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
